dio-project-fraude-detector/src/requirements.txt
streamlit
azure-ai-documentintelligence
azure-core
python-dotenv
azure-storage-blob

